+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [6.25.0]
-- update zoom.tools.now() for compatibility with Python 3.12+
+- update zoom.tools.now() for compatibility with Python >= 3.11
 - add admin theme
 - use passlib hash method instead of deprecated encrypt method
 - update tests and pipelines to remove deprecated calls